backend/cloud_functions/send_email/requirements.txt
sendgrid==6.11.0
google-cloud-firestore==2.13.1
functions-framework==3.5.0

